Conquest Automodchat + restructuring of if gate curly brackets
The bot now checks for Conquest in the set RP if the RP is not in amphy
when .setAt is false, i.e. the RP has not stated, when .sethost is
called.  If Conquest is picked up by the regexp, it will set modgchat
+.  Modchat is removed upon the use of .start, .rmhost and .endrp,
which all check that the RP is not in progress, in case it accidentally
removes modgchat.