Main Voices, RustyRP, RPPolls, and Legends.
Main voices are now able to use the commands they’re allowed to use in
main in amphy.

RustyRP patches.
- Don’t devoice “hosting voices” in rusty
- Don’t have void work in Rusty

Void formatting changes.  Void now works for regs, but PMs, void now
has a codlin like host and rp.

RPPoll is now a function.
Legends is now a function and links to the revised poll thread.