Implementing Battleless CQ as an official variant.
When CQ is voted, creates a poll, with the options: Regular CQ,
Battleless CQ, and Variant CQ.

Defaultdoc has been updated so that when CQ voted from the main RP
Poll, we set the rules docs for both regular, and battleless.  Then
after one or the other is voted, the respective hosting docs are set.